Create separate components for current output and capacity
